+++ src/command_modules/azure-cli-role/azure/cli/command_modules/role/_params.py
@@ -44,6 +44,8 @@ def load_arguments(self, _):
         c.argument('oauth2_allow_implicit_flow', arg_type=get_three_state_flag(), help='whether to allow implicit grant flow for OAuth2')
         c.argument('required_resource_accesses', type=validate_file_or_dict,
                    help="resource scopes and roles the application requires access to. Should be in manifest json format. See examples below for details")
+        c.argument('app_roles', type=validate_file_or_dict,
+                   help="declare the roles you want to associate with your application. Should be in manifest json format. See examples below for details")
         c.argument('native_app', arg_type=get_three_state_flag(), help="an application which can be installed on a user's device or computer")
         c.argument('credential_description', help="the description of the password")