+++ src/command_modules/azure-cli-container/HISTORY.rst
@@ -3,6 +3,12 @@
 Release History
 ===============
 
+* `container start/restart`: Added `--no-wait` argument.
+
+0.3.14
+++++++
+* Minor fixes.
+
 0.3.13
 ++++++
 * Adding 'az container start' command