+++ src/command_modules/azure-cli-hdinsight/HISTORY.rst
@@ -3,6 +3,11 @@
 Release History
 ===============
 
+0.3.1
++++++
+
+* `create`: added the `--storage-account-managed-identity` parameter to support ADLS Gen2 MSI.
+
 0.3.0
 +++++